End Game Menu Implemented
End Game menu is displayed if the player takes longer than the allowed
time in the TimeTrial Maze. WinMenu is no longer attached to a prefab.
Instead it is displayed the same way as the LoseMenu. When the player
reaches the last cell of the maze, the WinMenu just needs to be set to
active.